bugfixes
Code refactoring is done, and some anoying bugs chased down, the
readscan was adding an copy of an already exsisting system to wh links
instead of a reference, now found and fixed thanks to ekolis.  initial
system map moving is in, but not really working properly yet. system
placement is much better and is nolonger replacing the systems multiple
times. still need to fix so that when reset systems is uncheckd it
doesnt just place everythin in one spot. new winbuildexe done.