Persistant map location.
system icons are now persistent. if you move your systems around in the
map, then save the map, next time you load it and display it (as long as
the reset locs is not checked) the systems will appear as they were when
you saved it. fixed a small problem with the width of the systems, but
created another small problem with the width of the systems when notes
are active.